+++ selfdrive/ui/replay/route.h
@@ -9,6 +9,13 @@
 
 const QDir CACHE_DIR(util::getenv("COMMA_CACHE", "/tmp/comma_download_cache/").c_str());
 
+struct RouteIdentifier {
+  QString dongle_id;
+  QString timestamp;
+  int segment_id;
+  QString str;
+};
+
 struct SegmentFile {
   QString rlog;
   QString qlog;
@@ -20,18 +27,20 @@ struct SegmentFile {
 
 class Route {
 public:
-  Route(const QString &route, const QString &data_dir = {}) : route_(route), data_dir_(data_dir) {};
+  Route(const QString &route, const QString &data_dir = {});
   bool load();
-  inline const QString &name() const { return route_; };
+  inline const QString &name() const { return route_.str; }
+  inline const RouteIdentifier &identifier() const { return route_; }
   inline const std::map<int, SegmentFile> &segments() const { return segments_; }
   inline const SegmentFile &at(int n) { return segments_.at(n); }
+  static RouteIdentifier parseRoute(const QString &str);
 
 protected:
   bool loadFromLocal();
   bool loadFromServer();
   bool loadFromJson(const QString &json);
   void addFileToSegment(int seg_num, const QString &file);
-  QString route_;
+  RouteIdentifier route_;
   QString data_dir_;
   std::map<int, SegmentFile> segments_;
 };
